refactor: share helper between getBlockTextBefore and getBlockTextAfter
Both selectors had the same shape: pick a selection point, find the
focus text block at that point, compute the block's matching edge, and
return text between the point and the edge. Extracted the shared work
into a private `getBlockTextAroundPoint(snapshot, point, edge)` helper.
The two public selectors stay as named exports.
