+++ packages/editor/src/selectors/selector.get-text-before.ts
@@ -1,8 +1,6 @@
 import type {EditorSelector} from '../editor/editor-selector'
-import {getBlockStartPoint} from '../utils/util.get-block-start-point'
 import {getSelectionStartPoint} from '../utils/util.get-selection-start-point'
-import {getFocusTextBlock} from './selector.get-focus-text-block'
-import {getSelectionText} from './selector.get-selection-text'
+import {getBlockTextAroundPoint} from './selector.get-block-text-around-point'
 
 /**
  * @public
@@ -12,35 +10,9 @@ export const getBlockTextBefore: EditorSelector<string> = (snapshot) => {
     return ''
   }
 
-  const startPoint = getSelectionStartPoint(snapshot.context.selection)
-  const block = getFocusTextBlock({
-    ...snapshot,
-    context: {
-      ...snapshot.context,
-      selection: {
-        anchor: startPoint,
-        focus: startPoint,
-      },
-    },
-  })
-
-  if (!block) {
-    return ''
-  }
-
-  const startOfBlock = getBlockStartPoint({
-    context: snapshot.context,
-    block,
-  })
-
-  return getSelectionText({
-    ...snapshot,
-    context: {
-      ...snapshot.context,
-      selection: {
-        anchor: startOfBlock,
-        focus: startPoint,
-      },
-    },
-  })
+  return getBlockTextAroundPoint(
+    snapshot,
+    getSelectionStartPoint(snapshot.context.selection),
+    'start',
+  )
 }
